move TestDataUploadRepository to upload module
